chore: add eclipse warp routes to known app contexts
