+++ typescript/infra/config/environments/mainnet3/warp/warpIds.ts
@@ -1,0 +1,17 @@
+export enum WarpRouteIds {
+  Ancient8EthereumUSDC = 'USDC/ancient8-ethereum',
+  ArbitrumBaseBlastBscEthereumFraxtalLineaModeOptimismZircuitEZETH = 'EZETH/arbitrum-base-blast-bsc-ethereum-fraxtal-linea-mode-optimism-zircuit',
+  ArbitrumNeutronEclip = 'ECLIP/arbitrum-neutron',
+  ArbitrumNeutronTIA = 'TIA/arbitrum-neutron',
+  EclipseSolanaSOL = 'SOL/eclipsemainnet-solanamainnet',
+  EclipseSolanaWIF = 'WIF/eclipsemainnet-solanamainnet',
+  EthereumInevmUSDC = 'USDC/ethereum-inevm',
+  EthereumInevmUSDT = 'USDT/ethereum-inevm',
+  EthereumEclipseTETH = 'tETH/eclipsemainnet-ethereum',
+  EthereumEclipseUSDC = 'USDC/eclipsemainnet-ethereum-solanamainnet',
+  EthereumVictionETH = 'ETH/ethereum-viction',
+  EthereumVictionUSDC = 'USDC/ethereum-viction',
+  EthereumVictionUSDT = 'USDT/ethereum-viction',
+  InevmInjectiveINJ = 'INJ/inevm-injective',
+  MantapacificNeutronTIA = 'TIA/mantapacific-neutron',
+}
